+++ rendering/ProbabilitiesRenderer.cpp
@@ -2,7 +2,7 @@
 #include "Compositor.hpp"
 #include <iostream>
 
-#define COMPUTEINDEXOF(x, y, z) ( (x) * (ylength) * (zlength) + (y) * (zlength) + (z) )
+#define COMPUTEINDEXOF(x, y, z) (int)( (x) * (ylength) * (zlength) + (y) * (zlength) + (z) )
 std::vector<double> velocityMagnitudes;
 
 std::vector<double> ProbabilitiesRenderer::GetStartPoint()
@@ -62,10 +62,10 @@ void ProbabilitiesRenderer::PrepareGeometry(DataProvider* provider)
         probabilities.push_back(probDump);
 
         if ( provider->GetField(probName, &probabilities[i]) != 0)
-    	{
+        {
             std::cout << "ERROR<ProbabilitiesRenderer::PrepareGeometry>: " << probName << " Field could not be retrieved!" << std::endl;
             return;
-    	}
+        }
     }
 
     std::cout << "ProbabilitiesRenderer::PrepareGeometry -- processing " << (*points).size() << " points" << std::endl;
@@ -76,7 +76,7 @@ void ProbabilitiesRenderer::PrepareGeometry(DataProvider* provider)
     }
     if (this->totalAttributes > 0)
     {
-        for (int i = 0; i < this->totalAttributes; i++)
+        for (unsigned int i = 0; i < this->totalAttributes; i++)
         {
             delete(this->vertex_attrib_data[i]);
         }
@@ -96,64 +96,64 @@ void ProbabilitiesRenderer::PrepareGeometry(DataProvider* provider)
     double maxX = glm::min(xlength, endPoint[0]);
     double maxY = glm::min(ylength, endPoint[1]);
     double maxZ = glm::min(zlength, endPoint[2]);
-    this->totalVertices = (maxZ - minZ + 1) * (maxY - minY + 1) * (maxX - minX + 1);
+    this->totalVertices = (unsigned int)((maxZ - minZ + 1) * (maxY - minY + 1) * (maxX - minX + 1));
     this->vertex_buffer_data = new GLfloat[18 * ArrowGlyphSize * 3 * this->totalVertices]; // 3 floats per vertex
 
 
     int globalCounter = 0;
 
-    for (int i = minX; i <= maxX; i++)
+    for (int i = (int)minX; i <= (int)maxX; i++)
     {
-    	for (int j = minY; j <= maxY; j++)
+        for (int j = (int)minY; j <= (int)maxY; j++)
         {
-            for (int k = minZ; k <= maxZ; k++)
+            for (int k = (int)minZ; k <= (int)maxZ; k++)
             {
                 for(int l = 0; l < 19; l++)
                 {
                     if (l != 9)
                     {
-			            glm::mat4 M = glm::mat4(1.0f);
-			            M = glm::translate(M,  glm::vec3((points->at(COMPUTEINDEXOF(i, j, k)))[0],    // translation matrix to current location in dataset
+                        glm::mat4 M = glm::mat4(1.0f);
+                        M = glm::translate(M,  glm::vec3((points->at(COMPUTEINDEXOF(i, j, k)))[0],    // translation matrix to current location in dataset
                                          (points->at(COMPUTEINDEXOF(i, j, k)))[1],
                                          (points->at(COMPUTEINDEXOF(i, j, k)))[2]));
-			            glm::vec3 source_vec = glm::normalize(glm::vec3(0.0, 0.0, 1.0));                      // our current direction (all glyphs face +Z by default)
-			            glm::vec3 target_vec = glm::vec3(LATTICEVELOCITIES[l][0], LATTICEVELOCITIES[l][1], LATTICEVELOCITIES[l][2]); // vector facing direction we want to face
+                        glm::vec3 source_vec = glm::normalize(glm::vec3(0.0, 0.0, 1.0));                      // our current direction (all glyphs face +Z by default)
+                        glm::vec3 target_vec = glm::vec3(LATTICEVELOCITIES[l][0], LATTICEVELOCITIES[l][1], LATTICEVELOCITIES[l][2]); // vector facing direction we want to face
 
                         if (glm::length(target_vec) > 0.0)
-			            {
-			                target_vec = glm::normalize(target_vec);
+                        {
+                            target_vec = glm::normalize(target_vec);
                             glm::vec3 rot_axis = glm::cross(source_vec, target_vec);
 
-        					if ((rot_axis[0])*(rot_axis[0]) <= 0.0000001 && (rot_axis[1])*(rot_axis[1]) <= 0.0000001 && (rot_axis[2])*(rot_axis[2]) <= 0.0000001)
+                            if ((rot_axis[0])*(rot_axis[0]) <= 0.0000001 && (rot_axis[1])*(rot_axis[1]) <= 0.0000001 && (rot_axis[2])*(rot_axis[2]) <= 0.0000001)
                             {
-		        			    if (glm::dot(source_vec, target_vec) < 0)
+                                if (glm::dot(source_vec, target_vec) < 0)
                                 {
-				            		glm::vec3 temp = target_vec;
-            						temp[0] = temp[0] + 1.432342; temp[1] = temp[1] + 1.234235342; temp[2] = temp[2] + 1.1244325;
-			            			rot_axis = glm::cross(source_vec, temp);
-            						M = glm::rotate(M, 3.1415f, rot_axis);
-			        		    }
-					        }
-        					else
+                                    glm::vec3 temp = target_vec;
+                                    temp[0] = temp[0] + 1.432342f; temp[1] = temp[1] + 1.234235342f; temp[2] = temp[2] + 1.1244325f;
+                                    rot_axis = glm::cross(source_vec, temp);
+                                    M = glm::rotate(M, 3.1415f, rot_axis);
+                                }
+                            }
+                            else
                             {
-    			        	    float rot_angle = glm::acos(glm::dot(source_vec, target_vec));
-	    		        	    M = glm::rotate(M, rot_angle, rot_axis);                              // rotation matrix from (0,0,1) to velocity dir at this location
-		        			}
-			            }
+                                float rot_angle = glm::acos(glm::dot(source_vec, target_vec));
+                                M = glm::rotate(M, rot_angle, rot_axis);                              // rotation matrix from (0,0,1) to velocity dir at this location
+                            }
+                        }
 
-			            // Loop through the arrow skeleton
-			            for (int loopVarGlyphPts = 0; loopVarGlyphPts < ArrowGlyphSize * 3; loopVarGlyphPts += 3)
-			            {
-        					glm::vec4 glyphPointTemp;
-			                // get coords for current vertex
-		        			if (l == 2 || l == 6 || l == 9 || l == 10 || l == 12 || l == 16)
+                        // Loop through the arrow skeleton
+                        for (int loopVarGlyphPts = 0; loopVarGlyphPts < ArrowGlyphSize * 3; loopVarGlyphPts += 3)
+                        {
+                            glm::vec4 glyphPointTemp;
+                            // get coords for current vertex
+                            if (l == 2 || l == 6 || l == 9 || l == 10 || l == 12 || l == 16)
                             {
-				                glyphPointTemp = glm::vec4(
+                                glyphPointTemp = glm::vec4(
                                             ((probabilities[l])->at(COMPUTEINDEXOF(i, j, k)))[0]*VectorScale*g_arrow2d_vertex_buffer_data[loopVarGlyphPts+0],
                                             ((probabilities[l])->at(COMPUTEINDEXOF(i, j, k)))[0]*VectorScale*g_arrow2d_vertex_buffer_data[loopVarGlyphPts+1],
                                             ((probabilities[l])->at(COMPUTEINDEXOF(i, j, k)))[0]*VectorScale*g_arrow2d_vertex_buffer_data[loopVarGlyphPts+2],
                                             1.0);
-        					}
+                            }
                             else
                             {
                                 glyphPointTemp = glm::vec4(
@@ -170,10 +170,10 @@ void ProbabilitiesRenderer::PrepareGeometry(DataProvider* provider)
                             if (probability > maxProbability)
                                 maxProbability = probability;
 
-			                // store (x,y,z) components of current vertex*/
+                            // store (x,y,z) components of current vertex*/
                             velocityMagnitudes.push_back(probability);
-			                for (int loopVarComponents = 0; loopVarComponents < 3; loopVarComponents++)
-			                {
+                            for (int loopVarComponents = 0; loopVarComponents < 3; loopVarComponents++)
+                            {
                                 this->vertex_buffer_data[globalCounter] = glyphPointTemp[loopVarComponents];
                                 globalCounter++;
                             }
@@ -194,7 +194,7 @@ void ProbabilitiesRenderer::PrepareGeometry(DataProvider* provider)
     this->vertex_attrib_data[0] = new GLfloat[num_of_vertices]; // 1 velocity magnitude per *vertex*
     for (int i = 0; i < num_of_vertices; i++)
     {
-        this->vertex_attrib_data[0][i] = velocityMagnitudes[i]/maxProbability;
+        this->vertex_attrib_data[0][i] = (GLfloat)(velocityMagnitudes[i]/maxProbability);
     }
 
     GLuint vao, vbo, velocity_buf;
@@ -258,11 +258,11 @@ void ProbabilitiesRenderer::Draw(glm::mat4 MVP)
     // set shaders
     shaderProgram->enable();
     glUniformMatrix4fv(shaderProgram->getUniform("MVP"), 1, GL_FALSE, &MVP[0][0]);
-    glUniform1f(shaderProgram->getUniform("max_scalar"), this->maxGradientValue);
-    glUniform1f(shaderProgram->getUniform("min_scalar"), this->minGradientValue);
+    glUniform1f(shaderProgram->getUniform("max_scalar"), (GLfloat)this->maxGradientValue);
+    glUniform1f(shaderProgram->getUniform("min_scalar"), (GLfloat)this->minGradientValue);
     glUniform4fv(shaderProgram->getUniform("hotColor"), 1, this->maxColor);
     glUniform4fv(shaderProgram->getUniform("coldColor"), 1, this->minColor);
-    glUniform1f(shaderProgram->getUniform("bias"), this->bias);
+    glUniform1f(shaderProgram->getUniform("bias"), (GLfloat)this->bias);
     glUniform1i(shaderProgram->getUniform("interpolator"), this->interpolator);
     glBindVertexArray(this->VAO);
